Refactor character creation and GUI state management
- Moved character creation logic into a separate module (`character_creation.rs`) for better organization and maintainability.
- Updated `GuiState` to manage character creation state and integrate it with the main GUI rendering loop.
- Enhanced level-up selection handling by introducing `LevelUpSelectionProgress` to track progress for various level-up choices.
- Improved rendering of character attributes and level-up options in the GUI.
- Removed redundant ability score setting logic and replaced it with a more dynamic approach based on class defaults.
- Made `table_column` and rendering functions public for better accessibility in the GUI context.

+++ engine/src/components/class.rs
@@ -87,7 +87,7 @@ pub struct Class {
     pub hit_die: DieSize,
     pub hp_per_level: u8,
 
-    pub default_abilities: HashMap<Ability, i32>,
+    pub default_abilities: HashMap<Ability, u8>,
 
     /// Saving throw proficiencies granted at level 1 (e.g. STR + CON for Fighter)
     pub saving_throw_proficiencies: [Ability; 2],
@@ -102,7 +102,7 @@ impl Class {
         name: ClassName,
         hit_die: DieSize,
         hp_per_level: u8,
-        default_abilities: HashMap<Ability, i32>,
+        default_abilities: HashMap<Ability, u8>,
         saving_throw_proficiencies: [Ability; 2],
         subclass_level: u8,
         subclasses: HashMap<SubclassName, Subclass>,
